// src/app/models/room.ts
type CoverImage = {
  url: string;
};
export type Image = {
  _key: string;
  url: string;
};
type Amunity = {
  _key: string;
  amenity: string;
  icon: string;
};
type Slug = {
  _type: string;
  current: string;
};
export type Room = {
  _id: string;
  coverImage: CoverImage;
  description: string;
  dimension: string;
  discount: number;
  images: Image[];
  isBooked: boolean;
  isFeatured: boolean;
  name: string;
  numberOFBeds: number;
  offeredAmenities: Amunity[];
  price: number;
  slug: Slug;
  specialNote: string;
  type: string;
};

export type CreateBookingDto = {
  user: string;
  hotelRoom: string;
  checkinDate: string;
  checkoutDate: string;
  numberOfDays: number;
  adults: number;
  children: number;
  totalPrice: number;
  discount: number;
};